Move settings classes
